@@ -10,24 +10,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.0",
-    "@angular/common": "~13.1.0",
-    "@angular/compiler": "~13.1.0",
-    "@angular/core": "~13.1.0",
-    "@angular/forms": "~13.1.0",
-    "@angular/platform-browser": "~13.1.0",
-    "@angular/platform-browser-dynamic": "~13.1.0",
-    "@angular/router": "~13.1.0",
+    "@angular/animations": "^14.0.6",
+    "@angular/common": "^14.0.6",
+    "@angular/compiler": "^14.0.6",
+    "@angular/core": "^14.0.6",
+    "@angular/forms": "^14.0.6",
+    "@angular/platform-browser": "^14.0.6",
+    "@angular/platform-browser-dynamic": "^14.0.6",
+    "@angular/router": "^14.0.6",
     "angular-cli-ghpages": "^1.0.0",
     "bootstrap": "^5.1.3",
+    "ngx-toastr": "^15.0.0",
     "rxjs": "~7.4.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.6",
-    "@angular/cli": "~13.2.6",
-    "@angular/compiler-cli": "~13.1.0",
+    "@angular-devkit/build-angular": "^14.0.6",
+    "@angular/cli": "^14.0.6",
+    "@angular/compiler-cli": "^14.0.6",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.10.0",
@@ -36,6 +37,6 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.7.4"
   }
 }
