@@ -9,23 +9,28 @@ import { ProyectosComponent } from './components/proyectos/proyectos.component';
 import { SkillsComponent } from './components/skills/skills.component';
 import { HeaderComponent } from './components/header/header.component';
 import { PerfilfotoComponent } from './components/perfilfoto/perfilfoto.component';
-import {HttpClientModule } from '@angular/common/http';
+import {HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { ExperienciaComponent } from './components/experiencia/experiencia.component';
 import { EducacionComponent } from './components/educacion/educacion.component';
-import { LoginComponent } from './components/login/login.component';
+import { LoginComponent } from './components/auth/login.component';
 import { ButtonEditComponent } from './components/button-edit-persona/button-edit-persona.component';
 import { ButtonEditSkillsComponent } from './components/button-add-skills/button-add-skills.component';
-/* import {ObjToArrayPipe} from './objToArray.pipe'; */
+ 
 import { FormsModule } from '@angular/forms';
 import { ButtonAddProyectoComponent } from './components/button-add-proyecto/button-add-proyecto.component';
 
 import { ButtonAddEducacionComponent } from './components/button-add-educacion/button-add-educacion.component';
-import { DatearPipe } from './datear.pipe';
+ 
 import { FormularioEditComponent } from './components/formulario-list/formulario-list.component';
 import { ButtonAddExperienciaComponent } from './components/button-add-experiencia/button-add-experiencia.component';
 
+
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ToastrModule } from 'ngx-toastr';
+import { MenuComponent } from './components/menu/menu.component';
+import {   ProdInterceptorService } from './components/interceptors/prod-interceptor.service';
+ 
+
 
 @NgModule({
   declarations: [
@@ -39,22 +44,24 @@ import { ToastrModule } from 'ngx-toastr';
     ExperienciaComponent,
     EducacionComponent,
     LoginComponent, 
-    ButtonEditComponent, ButtonEditSkillsComponent,
-    /* ObjToArrayPipe, */
+    ButtonEditComponent, 
+    ButtonEditSkillsComponent,     
     ButtonAddProyectoComponent,     
-    ButtonAddEducacionComponent, DatearPipe, FormularioEditComponent, ButtonAddExperienciaComponent,
-     
-
-    
+    ButtonAddEducacionComponent,
+    FormularioEditComponent, 
+    ButtonAddExperienciaComponent, 
+    MenuComponent
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,    
     HttpClientModule,
     BrowserAnimationsModule,
-    ToastrModule.forRoot()
+    ToastrModule.forRoot(),
+    FormsModule
+     
   ],
-  providers: [LoginComponent],
+  providers: [LoginComponent,{provide: HTTP_INTERCEPTORS, useClass: ProdInterceptorService, multi: true} ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
