@@ -1,8 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { formulario } from 'src/app/model/formulario.model';
-import { LoginComponent } from '../login/login.component';
+import { LoginComponent } from '../auth/login.component';
 import{formularioService} from '../../service/formulario.service';
+import { ToastrService } from 'ngx-toastr';
+import { TokenService } from 'src/app/service/token.service';
 
 @Component({
   selector: 'app-formulario',
@@ -15,16 +17,29 @@ export class FormularioComponent   {
   mostrar:boolean=false;
   FomularioEnviadoOk:boolean=false;
   FomularioEnviadoBad:boolean=false;
-  constructor(private formularioServ: formularioService, private login: LoginComponent,private activatedRoute: ActivatedRoute,
-    private router: Router) { }
+  isLogged = false;
+  roles!: string[];
+  isAdmin = false;
+
+  constructor(private formularioServ: formularioService,   private toastr: ToastrService,  private tokenService: TokenService) { }
 
   ngOnInit(){
-    this.loginok= this.getLogin();
+     
+    if (this.tokenService.getToken()) {
+      this.isLogged = true;
+    } else {
+      this.isLogged = false;
+    }
+
+    this.roles = this.tokenService.getAuthorities();
+    this.roles.forEach(rol => {
+      if (rol === 'ROLE_ADMIN') {
+        this.isAdmin = true;
+      }
+    });
 
   } 
-public getLogin(){
-    return this.login.loginok();
-}
+ 
 
   addFormulario(){
 
@@ -48,12 +63,16 @@ public getLogin(){
 public newFormulario(formulario: formulario):void {
   this.formularioServ.createFormulario(formulario).subscribe(
     data => {this.FomularioEnviadoOk=true;
-       
+      this.toastr.success('Formulario enviado' , '', {
+        timeOut: 3000, positionClass: 'toast-top-center'
+      });
        
     },
     err => {          
      
-      this.FomularioEnviadoBad=true;
+      this.toastr.error('Error al enviar formulario' , '', {
+        timeOut: 3000, positionClass: 'toast-top-center'
+      });
     }
    );  
 }
