@@ -1,13 +1,13 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
-import { TokenService } from '../service/token.service';
+import { TokenService } from '../../service/token.service';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProdGuardService implements CanActivate {
 
-  realRol: string;
+  realRol!: string;
 
   constructor(
     private tokenService: TokenService,
@@ -15,7 +15,7 @@ export class ProdGuardService implements CanActivate {
   ) { }
 
   canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
-    const expectedRol = route.data.expectedRol;
+    const expectedRol = route.data['expectedRol'];
     const roles = this.tokenService.getAuthorities();
     this.realRol = 'user';
     roles.forEach(rol => {
