@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { TokenService } from '../service/token.service';
+import { TokenService } from 'src/app/service/token.service';
 
 @Injectable({
   providedIn: 'root'
@@ -13,11 +13,12 @@ export class ProdInterceptorService implements HttpInterceptor {
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     let intReq = req;
     const token = this.tokenService.getToken();
+    console.log("Token Interceptor: " + token);
     if (token != null) {
       intReq = req.clone({ headers: req.headers.set('Authorization', 'Bearer ' + token)});
     }
     return next.handle(intReq);
   }
 }
-
-export const interceptorProvider = [{provide: HTTP_INTERCEPTORS, useClass: ProdInterceptorService, multi: true}];
+console.log("Interceptor");
+//export const interceptorProvider = [{provide: HTTP_INTERCEPTORS, useClass: ProdInterceptorService, multi: true}];
