@@ -2,7 +2,9 @@ import { HttpErrorResponse } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { persona } from 'src/app/model/persona.model';
 import { PersonaService } from 'src/app/service/persona.service';
-import { LoginComponent } from '../login/login.component';
+import { TokenService } from 'src/app/service/token.service';
+import { LoginComponent } from '../auth/login.component';
+
 
 @Component({
   selector: 'app-perfilfoto',
@@ -11,46 +13,61 @@ import { LoginComponent } from '../login/login.component';
 })
 export class PerfilfotoComponent implements OnInit {
   public date: Date = new Date();
-  persona!: persona  
-    
-  
-  loginok!: boolean;
+  persona!: persona
 
-  constructor(private persoService: PersonaService, private login: LoginComponent )  { }
+  isLogged = false;
+  roles!: string[];
+  isAdmin = false;
+
+  constructor(private persoService: PersonaService, private tokenService: TokenService) { }
 
   ngOnInit(): void {
+
+
+    if (this.tokenService.getToken()) {
+      this.isLogged = true;
+    } else {
+      this.isLogged = false;
+    }
+
+    this.roles = this.tokenService.getAuthorities();
+    this.roles.forEach(rol => {
+      if (rol === 'ROLE_ADMIN') {
+        this.isAdmin = true;
+      }
+
+    });
     this.getPersona();
-    this.loginok= this.getLogin();
+    console.log("isLogged in Perfil: " + this.isLogged);
+    console.log("isAdmin in Perfil: " + this.isAdmin);
 
   }
 
-   
-public getPersona():void{  
-  this.persoService.getPersona().subscribe( response  => {this.persona=response} );       
-      
-}
 
-public updatePersona(p: persona):void{  
-  this.persoService.updatePersona(p).subscribe( response  => {response=this.persona;   });    
-  
-  
-} 
-wait(ms: number){
-  let start = new Date().getTime();
-  let end = start;
-  while(end < start + ms) {
-    end = new Date().getTime();
- }
-}
+  public getPersona(): void {
+    this.persoService.getPersona().subscribe(response => { this.persona = response });
 
-public reloadComponent(evento: boolean){
-  this.wait(500);
-  this.ngOnInit();  
-}
+  }
 
-public getLogin(){
-  return this.login.loginok();
-}
+  public updatePersona(p: persona): void {
+    this.persoService.updatePersona(p).subscribe(response => { response = this.persona; });
+
+
+  }
+  wait(ms: number) {
+    let start = new Date().getTime();
+    let end = start;
+    while (end < start + ms) {
+      end = new Date().getTime();
+    }
+  }
+
+  public reloadComponent(evento: boolean) {
+    /*  this.wait(500);
+     this.ngOnInit(); */
+  }
+
+
 
 
 }
