@@ -1,6 +1,6 @@
 export class JwtDTO {
-    token?: string;
-    type?: string;
-    nombreUsuario?: string;
-    authorities?: string[];
+    token!: string;
+    type!: string;
+    nombreUsuario!: string;
+    authorities!: string[];
 }
