@@ -4,10 +4,13 @@ const TOKEN_KEY = 'AuthToken';
 const USERNAME_KEY = 'AuthUserName';
 const AUTHORITIES_KEY = 'AuthAuthorities';
 
+
 @Injectable({
   providedIn: 'root'
 })
 export class TokenService {
+
+
 
   roles: Array<string> = [];
 
@@ -19,7 +22,13 @@ export class TokenService {
   }
 
   public getToken(): string {
-    return sessionStorage.getItem(TOKEN_KEY);
+    console.log("Toke;: " + sessionStorage.getItem(TOKEN_KEY));
+    const valor = sessionStorage.getItem(TOKEN_KEY);
+    if (valor) {
+      return valor;
+    } else {
+      return "";
+    }
   }
 
   public setUserName(userName: string): void {
@@ -28,7 +37,13 @@ export class TokenService {
   }
 
   public getUserName(): string {
-    return sessionStorage.getItem(USERNAME_KEY);
+
+    const valor = sessionStorage.getItem(USERNAME_KEY)
+    if (valor) {
+      return valor;
+    }
+    return "none";
+
   }
 
   public setAuthorities(authorities: string[]): void {
@@ -39,9 +54,16 @@ export class TokenService {
   public getAuthorities(): string[] {
     this.roles = [];
     if (sessionStorage.getItem(AUTHORITIES_KEY)) {
-      JSON.parse(sessionStorage.getItem(AUTHORITIES_KEY)).forEach(authority => {
-        this.roles.push(authority.authority);
-      });
+
+      const valor = sessionStorage.getItem(AUTHORITIES_KEY);
+      if (valor) {
+
+        JSON.parse(valor).forEach((authority: { authority: string; }) => {
+          this.roles.push(authority.authority);
+        });
+       
+      }
+
     }
     return this.roles;
   }
